Add parseFloat to post request

+++ src/apiCalls.js
@@ -19,7 +19,7 @@ export const fetchAllIngredients = () => {
 export const adjustUserPantry = (ingredientInfo, pantry) => {
   return fetch('http://localhost:3001/api/v1/users', {
     method: 'POST',
-    body: JSON.stringify({userID: parseInt(`${pantry.id}`), ingredientID: parseInt(`${ingredientInfo.id}`), ingredientModification: parseInt(`${ingredientInfo.amountNeeded}`)}),
+    body: JSON.stringify({userID: parseFloat(`${pantry.id}`), ingredientID: parseFloat(`${ingredientInfo.id}`), ingredientModification: parseFloat(`${ingredientInfo.amountNeeded}`)}),
     headers: {
       'Content-type': 'application/json'
     }
